requirements.txt
google-genai
flask
sift-stack-py